interim controller added for diesel log and vehicle alerts

+++ Model/VehicleAlerts.cs
@@ -12,39 +12,39 @@ public class VehicleAlerts
     [Key]
     public int Id { get; set; }
 
-    public int VehicleId { get; set; }
+    public short VehicleId { get; set; }
 
     [Column(TypeName = "decimal(18,2)")]
     public decimal Emi { get; set; }
     
     public DateOnly EmiDueDate { get; set; }
 
-    public int EmiDaysLeft { get; set; }
+    public short EmiDaysLeft { get; set; }
 
     [Required]
     public DateOnly InsuranceExpiry   { get; set; }
 
 
-    public int InsuranceDaysLeft { get; set; }
+    public short InsuranceDaysLeft { get; set; }
 
     [Required]
-    public int InsuranceStatus { get; set; }
+    public short InsuranceStatus { get; set; }
 
 
     [Required]
     public DateOnly PollutionExpiry { get; set; } 
 
 
-    public int PollutionDaysLeft { get; set; }
+    public short PollutionDaysLeft { get; set; }
 
 
     [Required]
-    public int PollutionStatus { get; set; }
+    public short PollutionStatus { get; set; }
 
     public string? FuelStationName { get; set; }
 
     public string? Remarks { get; set; }
-    
+
 
     [ForeignKey("VehicleId")]
     public required Vehicle Vehicle { get; set; } // Navigation property to Vehicle
